+++ examples/run_rest.py
@@ -26,13 +26,16 @@ async def run() -> None:
     # token can also be used with Tanjun but this may limit functionality.
     bot = hikari.RESTBot(loaded_config.bot_token, hikari.TokenType.BOT)
     database = impls.DatabaseImpl()
-    client = (
+    (
         # Passing True for declare_global_commands here instructs the client to
         # declare the slash commands within it which are marked as "global" during
         # the first startup.
         # A guild ID may also be passed here to instruct it to just declare the
         # global commands for that guild, this can be helpful for debug purposes.
-        tanjun.Client.from_rest_bot(bot, declare_global_commands=True)
+        #
+        # `bot_managed=True` must be passed here to indicate that the client should
+        # be automatically started when the REST bot starts.
+        tanjun.Client.from_rest_bot(bot, declare_global_commands=True, bot_managed=True)
         # Unlike a gateway bot bound client, only slash commands will be automatically
         # executed by a client that's bound to a rest bot.
         .load_modules("examples.slash_component")
@@ -41,18 +44,7 @@ async def run() -> None:
         # Here we use client callbacks to manage the database, STOPPING can also be used to stop it.
         .add_client_callback(tanjun.ClientCallbackNames.STARTING, database.connect)
     )
-    # Unlike with a gateway bot, for RESTBots hikari has no lifetime event
-    # dispatch which can be used to implicitly startup and close the Tanjun
-    # client. Instead, we must directly startup and close Tanjun.
-    await bot.start()
-
-    # Note that starting a Tanjun client before the relevant bot may lead
-    # to erroneous behaviour as it won't be able to make any requests.
-
-    # While this example uses the client as a context manager to implicitly start
-    # and close it, the `open` and `close` methods can alternatively be used to the same effect.
-    async with client:
-        await bot.join()  # This waits until the bot is closed before closing Tanjun by exiting the context manager.
+    bot.run()
 
 
 def main():